refactor/ merge TextCopy and ImageCopy to CopiedItem
